Split app.py low-level functions to utils/app_utils.py ?
