+++ src/app.py
@@ -18,10 +18,7 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 
-import copy
-import glob
 import os
-import re
 import shutil
 import subprocess
 import tempfile
@@ -31,19 +28,12 @@ from pathlib import Path
 from typing import (
     TYPE_CHECKING,
     Any,
-    Dict,
-    Iterator,
-    List,
     Literal,
-    Optional,
-    Required,
     Tuple,
-    TypedDict,
     Union,
     cast,
 )
 
-import yaml
 from moulinette import Moulinette, m18n
 from moulinette.utils.filesystem import (
     chmod,
@@ -51,13 +41,10 @@ from moulinette.utils.filesystem import (
     cp,
     read_file,
     read_json,
-    read_toml,
     rm,
     write_to_file,
     write_to_json,
 )
-from moulinette.utils.process import check_output, run_commands
-from packaging import version
 
 from .app_catalog import (  # noqa
     APPS_CATALOG_LOGOS,
@@ -68,15 +55,36 @@ from .app_catalog import (  # noqa
 from .log import OperationLogger, is_flash_unit_operation, is_unit_operation
 from .utils.error import YunohostError, YunohostValidationError
 from .utils.i18n import _value_for_locale
-from .utils.system import (
-    binary_to_human,
-    debian_version,
-    dpkg_is_broken,
-    free_space_in_directory,
-    get_ynh_package_version,
-    human_to_binary,
-    ram_available,
-    system_arch,
+from .utils.system import free_space_in_directory
+from .utils.app_utils import (
+    _get_app_settings,
+    _set_app_settings,
+    _parse_app_version,
+    _get_manifest_of_app,
+    _hydrate_app_template,
+    _set_default_ask_questions,
+    _extract_app,
+    _is_installed,
+    _assert_is_installed,
+    _installed_apps,
+    _check_manifest_requirements,
+    _guess_webapp_path_requirement,
+    _validate_webpath_requirement,
+    _assert_no_conflicting_apps,
+    _make_environment_for_app_script,
+    _parse_app_instance_name,
+    _next_instance_number_for_app,
+    _make_tmp_workdir_for_app,
+    _assert_system_is_sane_for_app,
+    _notification_is_dismissed,
+    _filter_and_hydrate_notifications,
+    _display_notifications,
+    _confirm_app_install,
+    _ask_confirmation,
+    AppInfo,
+    AppManifest,
+    APPS_SETTING_PATH,
+    APP_TMP_WORKDIRS,
 )
 
 if TYPE_CHECKING:
@@ -90,17 +98,7 @@ if TYPE_CHECKING:
 else:
     logger = getLogger("yunohost.app")
 
-APPS_SETTING_PATH = "/etc/yunohost/apps/"
-APP_TMP_WORKDIRS = "/var/cache/yunohost/app_tmp_work_dirs"
 PORTAL_SETTINGS_DIR = "/etc/yunohost/portal"
-
-re_app_instance_name = re.compile(
-    r"^(?P<appid>[\w-]+?)(__(?P<appinstancenb>[1-9][0-9]*))?$"
-)
-
-APP_REPO_URL = re.compile(
-    r"^https://[a-zA-Z0-9-_.]+/[a-zA-Z0-9-_./~]+/[a-zA-Z0-9-_.]+_ynh(/?(-/)?(tree|src/(branch|tag|commit))/[a-zA-Z0-9-_.]+)?(\.git)?/?$"
-)
 
 APP_FILES_TO_COPY = [
     "manifest.json",
@@ -113,34 +111,6 @@ APP_FILES_TO_COPY = [
     "hooks",
     "doc",
 ]
-
-AppManifest = dict[str, Any]
-
-
-class AppInfo(TypedDict, total=False):
-    id: Required[str]
-    name: Required[str]
-    description: Required[str]
-    version: Required[str]
-    domain_path: str
-    logo: str | None
-    upgradable: Literal["yes", "no", "url_required", "bad_quality"]
-    current_version: str | None
-    new_version: str | None
-    settings: dict[str, Any]
-    setting_path: str
-    manifest: AppManifest
-    from_catalog: dict[str, Any]
-    is_webapp: bool
-    is_default: bool
-    supports_change_url: bool
-    supports_backup_restore: bool
-    supports_multi_instance: bool
-    supports_config_panel: bool
-    supports_purge: bool
-    permissions: dict[str, Any]
-    label: str
-    notifications: dict[str, dict[str, str]]
 
 
 def app_list(
@@ -471,11 +441,7 @@ def app_change_url(
     from .service import service_reload_or_restart
     from .utils.form import DomainOption, WebPathOption
 
-    installed = _is_installed(app)
-    if not installed:
-        raise YunohostValidationError(
-            "app_not_installed", app=app, all_apps=_get_all_installed_apps_id()
-        )
+    _assert_is_installed(app)
 
     if not os.path.exists(
         os.path.join(APPS_SETTING_PATH, app, "scripts", "change_url")
@@ -1074,26 +1040,6 @@ def app_manifest(app: str, with_screenshot: bool = False) -> AppManifest:
             manifest["requirements"][name] = "ok" if passed else m18n.n(err, **values)
 
     return manifest
-
-
-def _confirm_app_install(app: str, force: bool = False) -> None:
-    # Ignore if there's nothing for confirm (good quality app), if --force is used
-    # or if request on the API (confirm already implemented on the API side)
-    if force or Moulinette.interface.type == "api":
-        return
-
-    quality = _app_quality(app)
-    if quality == "success":
-        return
-
-    # i18n: confirm_app_install_warning
-    # i18n: confirm_app_install_danger
-    # i18n: confirm_app_install_thirdparty
-
-    if quality in ["danger", "thirdparty"]:
-        _ask_confirmation("confirm_app_install_" + quality, kind="hard")
-    else:
-        _ask_confirmation("confirm_app_install_" + quality, kind="soft")
 
 
 @is_unit_operation()
@@ -1919,12 +1865,7 @@ def app_ssowatconf() -> None:
 
 @is_flash_unit_operation()
 def app_change_label(app: str, new_label: str) -> None:
-    installed = _is_installed(app)
-    if not installed:
-        raise YunohostValidationError(
-            "app_not_installed", app=app, all_apps=_get_all_installed_apps_id()
-        )
-
+    _assert_is_installed(app)
     app_setting(app, "label", new_label)
 
     # FIXME: we kinda have redundant stuff between the label key on the main perm, and the label key at top level ...
@@ -2399,1212 +2340,6 @@ ynh_app_config_run $1
     return AppConfigPanel, AppCoreConfigPanel
 
 
-app_settings_cache: Dict[str, Dict[str, Any]] = {}
-app_settings_cache_timestamp: Dict[str, float] = {}
-
-
-def _get_app_settings(app: str) -> Dict[str, Any]:
-    """
-    Get settings of an installed app
-
-    Keyword arguments:
-        app -- The app id (like nextcloud__2)
-
-    """
-    _assert_is_installed(app)
-
-    app_setting_path = os.path.join(APPS_SETTING_PATH, app, "settings.yml")
-    app_setting_timestamp = os.path.getmtime(app_setting_path)
-
-    # perf: app settings are cached using the settings.yml's modification date,
-    # such that we don't have to worry too much about calling this function
-    # too many times (because ultimately parsing yml is not free)
-    if app_settings_cache_timestamp.get(app) == app_setting_timestamp:
-        return app_settings_cache[app].copy()
-
-    try:
-        with open(app_setting_path) as f:
-            settings = yaml.safe_load(f) or {}
-        # If label contains unicode char, this may later trigger issues when building strings...
-        # FIXME: this should be propagated to read_yaml so that this fix applies everywhere I think...
-        settings = {k: v for k, v in settings.items()}
-
-        # App settings should never be empty, there should always be at least some standard, internal keys like id, install_time etc.
-        # Otherwise, this probably means that the app settings disappeared somehow...
-        if not settings:
-            logger.error(
-                f"It looks like settings.yml for {app} is empty ... This should not happen ..."
-            )
-            logger.error(m18n.n("app_not_correctly_installed", app=app))
-            return {}
-
-        # Make the app id available as $app too
-        settings["app"] = app
-
-        # FIXME: it's not clear why this code exists... Shouldn't we hard-define 'id' as $app ...?
-        if app != settings["id"]:
-            return {}
-
-        # Cache the settings
-        app_settings_cache[app] = settings.copy()
-        app_settings_cache_timestamp[app] = app_setting_timestamp
-
-        return settings
-    except (IOError, TypeError, KeyError):
-        logger.error(m18n.n("app_not_correctly_installed", app=app))
-    return {}
-
-
-def _set_app_settings(app: str, settings: dict[str, Any]) -> None:
-    """
-    Set settings of an app
-
-    Keyword arguments:
-        app_id -- The app id (like nextcloud__2)
-        settings -- Dict with app settings
-
-    """
-    with open(os.path.join(APPS_SETTING_PATH, app, "settings.yml"), "w") as f:
-        yaml.safe_dump(settings, f, default_flow_style=False)
-
-    if app in app_settings_cache_timestamp:
-        del app_settings_cache_timestamp[app]
-    if app in app_settings_cache:
-        del app_settings_cache[app]
-
-
-def _parse_app_version(v: str) -> tuple[version.Version, int]:
-    if v in ["?", "-"]:
-        return (version.parse("0"), 0)
-
-    try:
-        if "~" in v:
-            return (
-                version.parse(v.split("~")[0]),
-                int(v.split("~")[1].replace("ynh", "")),
-            )
-        else:
-            return (version.parse(v), 0)
-    except Exception as e:
-        raise YunohostError(f"Failed to parse app version '{v}' : {e}", raw_msg=True)
-
-
-def _get_manifest_of_app(path: str) -> AppManifest:
-    "Get app manifest stored in json or in toml"
-
-    # sample data to get an idea of what is going on
-    # this toml extract:
-    #
-    # license = "free"
-    # url = "https://example.com"
-    # multi_instance = true
-    # version = "1.0~ynh1"
-    # packaging_format = 1
-    # services = ["nginx", "php7.0-fpm", "mysql"]
-    # id = "ynhexample"
-    # name = "YunoHost example app"
-    #
-    # [requirements]
-    # yunohost = ">= 3.5"
-    #
-    # [maintainer]
-    # url = "http://example.com"
-    # name = "John doe"
-    # email = "john.doe@example.com"
-    #
-    # [description]
-    # fr = "Exemple de package d'application pour YunoHost."
-    # en = "Example package for YunoHost application."
-    #
-    # [arguments]
-    #     [arguments.install.domain]
-    #     type = "domain"
-    #     example = "example.com"
-    #         [arguments.install.domain.ask]
-    #         fr = "Choisissez un nom de domaine pour ynhexample"
-    #         en = "Choose a domain name for ynhexample"
-    #
-    # will be parsed into this:
-    #
-    # OrderedDict([(u'license', u'free'),
-    #              (u'url', u'https://example.com'),
-    #              (u'multi_instance', True),
-    #              (u'version', u'1.0~ynh1'),
-    #              (u'packaging_format', 1),
-    #              (u'services', [u'nginx', u'php7.0-fpm', u'mysql']),
-    #              (u'id', u'ynhexample'),
-    #              (u'name', u'YunoHost example app'),
-    #              (u'requirements', OrderedDict([(u'yunohost', u'>= 3.5')])),
-    #              (u'maintainer',
-    #               OrderedDict([(u'url', u'http://example.com'),
-    #                            (u'name', u'John doe'),
-    #                            (u'email', u'john.doe@example.com')])),
-    #              (u'description',
-    #               OrderedDict([(u'fr',
-    #                             u"Exemple de package d'application pour YunoHost."),
-    #                            (u'en',
-    #                             u'Example package for YunoHost application.')])),
-    #              (u'arguments',
-    #               OrderedDict([(u'install',
-    #                             OrderedDict([(u'domain',
-    #                                           OrderedDict([(u'type', u'domain'),
-    #                                                        (u'example',
-    #                                                         u'example.com'),
-    #                                                        (u'ask',
-    #                                                         OrderedDict([(u'fr',
-    #                                                                       u'Choisissez un nom de domaine pour ynhexample'),
-    #                                                                      (u'en',
-    #                                                                       u'Choose a domain name for ynhexample')]))])),
-    #
-    # and needs to be converted into this:
-    #
-    # {
-    #     "name": "YunoHost example app",
-    #     "id": "ynhexample",
-    #     "packaging_format": 1,
-    #     "description": {
-    #     ¦   "en": "Example package for YunoHost application.",
-    #     ¦   "fr": "Exemple de package d’application pour YunoHost."
-    #     },
-    #     "version": "1.0~ynh1",
-    #     "url": "https://example.com",
-    #     "license": "free",
-    #     "maintainer": {
-    #     ¦   "name": "John doe",
-    #     ¦   "email": "john.doe@example.com",
-    #     ¦   "url": "http://example.com"
-    #     },
-    #     "requirements": {
-    #     ¦   "yunohost": ">= 3.5"
-    #     },
-    #     "multi_instance": true,
-    #     "services": [
-    #     ¦   "nginx",
-    #     ¦   "php7.0-fpm",
-    #     ¦   "mysql"
-    #     ],
-    #     "arguments": {
-    #     ¦   "install" : [
-    #     ¦   ¦   {
-    #     ¦   ¦   ¦   "name": "domain",
-    #     ¦   ¦   ¦   "type": "domain",
-    #     ¦   ¦   ¦   "ask": {
-    #     ¦   ¦   ¦   ¦   "en": "Choose a domain name for ynhexample",
-    #     ¦   ¦   ¦   ¦   "fr": "Choisissez un nom de domaine pour ynhexample"
-    #     ¦   ¦   ¦   },
-    #     ¦   ¦   ¦   "example": "example.com"
-    #     ¦   ¦   },
-
-    if os.path.exists(os.path.join(path, "manifest.toml")):
-        manifest = read_toml(os.path.join(path, "manifest.toml"))
-    elif os.path.exists(os.path.join(path, "manifest.json")):
-        manifest = read_json(os.path.join(path, "manifest.json"))
-    else:
-        raise YunohostError(
-            f"There doesn't seem to be any manifest file in {path} ... It looks like an app was not correctly installed/removed.",
-            raw_msg=True,
-        )
-
-    manifest["packaging_format"] = float(
-        str(manifest.get("packaging_format", "")).strip() or "0"
-    )
-
-    if manifest["packaging_format"] < 2:
-        manifest = _convert_v1_manifest_to_v2(manifest)
-
-    manifest["install"] = _set_default_ask_questions(manifest.get("install", {}))
-    manifest["doc"], manifest["notifications"] = _parse_app_doc_and_notifications(path)
-
-    return manifest
-
-
-def _parse_app_doc_and_notifications(path: str):
-    doc: dict[str, dict[str, str]] = {}
-    notification_names = ["PRE_INSTALL", "POST_INSTALL", "PRE_UPGRADE", "POST_UPGRADE"]
-
-    for filepath in glob.glob(os.path.join(path, "doc") + "/*.md"):
-        # to be improved : [a-z]{2,3} is a clumsy way of parsing the
-        # lang code ... some lang code are more complex that this é_è
-        m = re.match("([A-Z]*)(_[a-z]{2,3})?.md", filepath.split("/")[-1])
-
-        if not m:
-            # FIXME: shall we display a warning ? idk
-            continue
-
-        pagename, lang = m.groups()
-
-        if pagename in notification_names:
-            continue
-
-        lang = lang.strip("_") if lang else "en"
-
-        if pagename not in doc:
-            doc[pagename] = {}
-
-        try:
-            doc[pagename][lang] = read_file(filepath).strip()
-        except Exception as e:
-            logger.error(e)
-            continue
-
-    notifications: dict[str, dict[str, dict[str, str]]] = {}
-
-    for step in notification_names:
-        notifications[step] = {}
-        for filepath in glob.glob(os.path.join(path, "doc", f"{step}*.md")):
-            m = re.match(step + "(_[a-z]{2,3})?.md", filepath.split("/")[-1])
-            if not m:
-                continue
-            pagename = "main"
-            lang = m.groups()[0].strip("_") if m.groups()[0] else "en"
-            if pagename not in notifications[step]:
-                notifications[step][pagename] = {}
-            try:
-                notifications[step][pagename][lang] = read_file(filepath).strip()
-            except Exception as e:
-                logger.error(e)
-                continue
-
-        for filepath in glob.glob(os.path.join(path, "doc", f"{step}.d") + "/*.md"):
-            m = re.match(
-                r"([A-Za-z0-9\.\~]*)(_[a-z]{2,3})?.md", filepath.split("/")[-1]
-            )
-            if not m:
-                continue
-            pagename, lang = m.groups()
-            lang = lang.strip("_") if lang else "en"
-            if pagename not in notifications[step]:
-                notifications[step][pagename] = {}
-
-            try:
-                notifications[step][pagename][lang] = read_file(filepath).strip()
-            except Exception as e:
-                logger.error(e)
-                continue
-
-    return doc, notifications
-
-
-def _hydrate_app_template(template: str, data: dict[str, Any]):
-    # Apply jinja for stuff like {% if .. %} blocks,
-    # but only if there's indeed an if block (to try to reduce overhead or idk)
-    if "{%" in template:
-        from jinja2 import Template
-
-        template = Template(template).render(**data)
-
-    stuff_to_replace = set(re.findall(r"__[A-Z0-9]+?[A-Z0-9_]*?[A-Z0-9]*?__", template))
-
-    for stuff in stuff_to_replace:
-        varname = stuff.strip("_").lower()
-
-        if varname in data:
-            template = template.replace(stuff, str(data[varname]))
-
-    return template.strip()
-
-
-def _convert_v1_manifest_to_v2(manifest: dict[str, Any]) -> AppManifest:
-    manifest = copy.deepcopy(manifest)
-
-    if "upstream" not in manifest:
-        manifest["upstream"] = {}
-
-    if "license" in manifest and "license" not in manifest["upstream"]:
-        manifest["upstream"]["license"] = manifest["license"]
-
-    if "url" in manifest and "website" not in manifest["upstream"]:
-        manifest["upstream"]["website"] = manifest["url"]
-
-    manifest["integration"] = {
-        "yunohost": manifest.get("requirements", {})
-        .get("yunohost", "")
-        .replace(">", "")
-        .replace("=", "")
-        .replace(" ", ""),
-        "architectures": "?",
-        "multi_instance": manifest.get("multi_instance", False),
-        "ldap": "?",
-        "sso": "?",
-        "disk": "?",
-        "ram": {"build": "?", "runtime": "?"},
-    }
-
-    maintainers = manifest.get("maintainer", {})
-    if isinstance(maintainers, list):
-        maintainers = [m["name"] for m in maintainers]
-    else:
-        maintainers = [maintainers["name"]] if maintainers.get("name") else []
-
-    manifest["maintainers"] = maintainers
-
-    install_questions = manifest["arguments"]["install"]
-
-    manifest["install"] = {}
-    for question in install_questions:
-        name = question.pop("name")
-        if "ask" in question and name in [
-            "domain",
-            "path",
-            "admin",
-            "is_public",
-            "password",
-        ]:
-            question.pop("ask")
-        if question.get("example") and question.get("type") in [
-            "domain",
-            "path",
-            "user",
-            "boolean",
-            "password",
-        ]:
-            question.pop("example")
-
-        manifest["install"][name] = question
-
-    manifest["resources"] = {"system_user": {}, "install_dir": {"alias": "final_path"}}
-
-    keys_to_keep = [
-        "packaging_format",
-        "id",
-        "name",
-        "description",
-        "version",
-        "maintainers",
-        "upstream",
-        "integration",
-        "install",
-        "resources",
-    ]
-
-    keys_to_del = [key for key in manifest.keys() if key not in keys_to_keep]
-    for key in keys_to_del:
-        del manifest[key]
-
-    return manifest
-
-
-def _set_default_ask_questions(questions: dict[str, Any], script_name: str = "install"):
-    # arguments is something like
-    # { "domain":
-    #       {
-    #         "type": "domain",
-    #         ....
-    #       },
-    #    "path": {
-    #         "type": "path",
-    #         ...
-    #       },
-    #       ...
-    # }
-
-    # We set a default for any question with these matching (type, name)
-    #                           type       namei
-    # N.B. : this is only for install script ... should be reworked for other
-    # scripts if we supports args for other scripts in the future...
-    questions_with_default = [
-        ("domain", "domain"),  # i18n: app_manifest_install_ask_domain
-        ("path", "path"),  # i18n: app_manifest_install_ask_path
-        ("password", "password"),  # i18n: app_manifest_install_ask_password
-        ("user", "admin"),  # i18n: app_manifest_install_ask_admin
-        ("boolean", "is_public"),  # i18n: app_manifest_install_ask_is_public
-        (
-            "group",
-            "init_main_permission",
-        ),  # i18n: app_manifest_install_ask_init_main_permission
-        (
-            "group",
-            "init_admin_permission",
-        ),  # i18n: app_manifest_install_ask_init_admin_permission
-    ]
-
-    for question_id, question in questions.items():
-        question["id"] = question_id
-
-        # If this question corresponds to a question with default ask message...
-        if any(
-            (question.get("type"), question["id"]) == question_with_default
-            for question_with_default in questions_with_default
-        ):
-            # The key is for example "app_manifest_install_ask_domain"
-            question["ask"] = m18n.n(f"app_manifest_{script_name}_ask_{question['id']}")
-
-            # Also it in fact doesn't make sense for any of those questions to have an example value nor a default value...
-            if question.get("type") in ["domain", "user", "password"]:
-                if "example" in question:
-                    del question["example"]
-                if "default" in question:
-                    del question["default"]
-
-    return questions
-
-
-def _is_app_repo_url(string: str) -> bool:
-    string = string.strip()
-
-    # Dummy test for ssh-based stuff ... should probably be improved somehow
-    if "@" in string:
-        return True
-
-    return bool(APP_REPO_URL.match(string))
-
-
-def _app_quality(src: str) -> str:
-    """
-    app may in fact be an app name, an url, or a path
-    """
-
-    raw_app_catalog = _load_apps_catalog()["apps"]
-    if src in raw_app_catalog or _is_app_repo_url(src):
-        # If we got an app name directly (e.g. just "wordpress"), we gonna test this name
-        if src in raw_app_catalog:
-            app_name_to_test = src
-        # If we got an url like "https://github.com/foo/bar_ynh, we want to
-        # extract "bar" and test if we know this app
-        elif ("http://" in src) or ("https://" in src):
-            app_name_to_test = src.strip("/").split("/")[-1].replace("_ynh", "")
-        else:
-            # FIXME : watdo if '@' in app ?
-            return "thirdparty"
-
-        if app_name_to_test in raw_app_catalog:
-            state = raw_app_catalog[app_name_to_test].get("state", "notworking")
-            level = raw_app_catalog[app_name_to_test].get("level", None)
-            if state in ["working", "validated"]:
-                if isinstance(level, int) and level >= 5:
-                    return "success"
-                elif isinstance(level, int) and level > 0:
-                    return "warning"
-            return "danger"
-        else:
-            return "thirdparty"
-
-    elif os.path.exists(src):
-        return "thirdparty"
-    else:
-        if "http://" in src or "https://" in src:
-            logger.error(
-                f"{src} is not a valid app url: app url are expected to look like https://domain.tld/path/to/repo_ynh"
-            )
-        raise YunohostValidationError("app_unknown")
-
-
-def _extract_app(src: str) -> Tuple[Dict, str]:
-    """
-    src may be an app name, an url, or a path
-    """
-
-    raw_app_catalog = _load_apps_catalog()["apps"]
-
-    # App is an appname in the catalog
-    if src in raw_app_catalog:
-        if "git" not in raw_app_catalog[src]:
-            raise YunohostValidationError("app_unsupported_remote_type")
-
-        app_info = raw_app_catalog[src]
-        url = app_info["git"]["url"]
-        branch = app_info["git"]["branch"]
-        revision = str(app_info["git"]["revision"])
-        return _extract_app_from_gitrepo(
-            url, branch=branch, revision=revision, app_info=app_info
-        )
-    # App is a git repo url
-    elif _is_app_repo_url(src):
-        url = src.strip().strip("/")
-        # gitlab urls may look like 'https://domain/org/group/repo/-/tree/testing'
-        # compated to github urls looking like 'https://domain/org/repo/tree/testing'
-        if "/-/" in url:
-            url = url.replace("/-/", "/")
-        if "/tree/" in url:
-            url, branch = url.split("/tree/", 1)
-        else:
-            branch = None
-        return _extract_app_from_gitrepo(url, branch=branch)
-    # App is a local folder
-    elif os.path.exists(src):
-        return _extract_app_from_folder(src)
-    else:
-        if "http://" in src or "https://" in src:
-            logger.error(
-                f"{src} is not a valid app url: app url are expected to look like https://domain.tld/path/to/repo_ynh"
-            )
-        raise YunohostValidationError("app_unknown")
-
-
-def _extract_app_from_folder(path: str) -> Tuple[Dict, str]:
-    """
-    Unzip / untar / copy application tarball or directory to a tmp work directory
-
-    Keyword arguments:
-        path -- Path of the tarball or directory
-    """
-    logger.debug(m18n.n("extracting"))
-
-    path = os.path.abspath(path)
-
-    extracted_app_folder = _make_tmp_workdir_for_app()
-
-    if os.path.isdir(path):
-        shutil.rmtree(extracted_app_folder)
-        if path[-1] != "/":
-            path = path + "/"
-        cp(path, extracted_app_folder, recursive=True)
-        # Change the last edit time which is used in _make_tmp_workdir_for_app
-        # to cleanup old dir ... otherwise it may end up being incorrectly removed
-        # at the end of the safety-backup-before-upgrade :/
-        os.system(f"touch {extracted_app_folder}")
-    else:
-        try:
-            shutil.unpack_archive(path, extracted_app_folder)
-        except Exception:
-            raise YunohostError("app_extraction_failed")
-
-    try:
-        if len(os.listdir(extracted_app_folder)) == 1:
-            for folder in os.listdir(extracted_app_folder):
-                extracted_app_folder = extracted_app_folder + "/" + folder
-    except IOError:
-        raise YunohostError("app_install_files_invalid")
-
-    manifest = _get_manifest_of_app(extracted_app_folder)
-    manifest["lastUpdate"] = int(time.time())
-
-    logger.debug(m18n.n("done"))
-
-    manifest["remote"] = {"type": "file", "path": path}
-    manifest["quality"] = {"level": -1, "state": "thirdparty"}
-    manifest["antifeatures"] = []
-    manifest["potential_alternative_to"] = []
-
-    return manifest, extracted_app_folder
-
-
-def _extract_app_from_gitrepo(
-    url: str, branch: Optional[str] = None, revision: str = "HEAD", app_info: Dict = {}
-) -> Tuple[Dict, str]:
-    logger.debug("Checking default branch")
-
-    try:
-        git_ls_remote = check_output(
-            ["git", "ls-remote", "--symref", url, "HEAD"],
-            env={"GIT_TERMINAL_PROMPT": "0", "LC_ALL": "C"},
-            shell=False,
-        )
-    except Exception as e:
-        logger.error(str(e))
-        raise YunohostError("app_sources_fetch_failed")
-
-    if not branch:
-        default_branch = None
-        try:
-            for line in git_ls_remote.split("\n"):
-                # Look for the line formated like :
-                # ref: refs/heads/master	HEAD
-                if "ref: refs/heads/" in line:
-                    line = line.replace("/", " ").replace("\t", " ")
-                    default_branch = line.split()[3]
-        except Exception:
-            pass
-
-        if not default_branch:
-            logger.warning("Failed to parse default branch, trying 'main'")
-            branch = "main"
-        else:
-            if default_branch in ["testing", "dev"]:
-                logger.warning(
-                    f"Trying 'master' branch instead of default '{default_branch}'"
-                )
-                branch = "master"
-            else:
-                branch = default_branch
-
-    logger.debug(m18n.n("downloading"))
-
-    extracted_app_folder = _make_tmp_workdir_for_app()
-
-    # Download only this commit
-    try:
-        # We don't use git clone because, git clone can't download
-        # a specific revision only
-        ref = branch if revision == "HEAD" else revision
-        run_commands([["git", "init", extracted_app_folder]], shell=False)
-        run_commands(
-            [
-                ["git", "remote", "add", "origin", url],
-                ["git", "fetch", "--depth=1", "origin", ref],
-                ["git", "reset", "--hard", "FETCH_HEAD"],
-            ],
-            cwd=extracted_app_folder,
-            shell=False,
-        )
-    except subprocess.CalledProcessError:
-        raise YunohostError("app_sources_fetch_failed")
-    else:
-        logger.debug(m18n.n("done"))
-
-    manifest = _get_manifest_of_app(extracted_app_folder)
-
-    # Store remote repository info into the returned manifest
-    manifest["remote"] = {"type": "git", "url": url, "branch": branch}
-    if revision == "HEAD":
-        try:
-            # Get git last commit hash
-            cmd = f"git ls-remote --exit-code {url} {branch} | awk '{{print $1}}'"
-            manifest["remote"]["revision"] = check_output(cmd)
-        except Exception as e:
-            logger.warning(f"cannot get last commit hash because: {e}")
-    else:
-        manifest["remote"]["revision"] = revision
-        manifest["lastUpdate"] = app_info.get("lastUpdate")
-
-    manifest["quality"] = {
-        "level": app_info.get("level", -1),
-        "state": app_info.get("state", "thirdparty"),
-    }
-    manifest["antifeatures"] = app_info.get("antifeatures", [])
-    manifest["potential_alternative_to"] = app_info.get("potential_alternative_to", [])
-
-    return manifest, extracted_app_folder
-
-
-def _list_upgradable_apps() -> list[AppInfo]:
-    upgradable_apps = list(app_list(upgradable=True)["apps"])
-
-    # Retrieve next manifest pre_upgrade notifications
-    for app in upgradable_apps:
-        absolute_app_name, _ = _parse_app_instance_name(app["id"])
-        manifest, extracted_app_folder = _extract_app(absolute_app_name)
-        app["notifications"] = {}
-        if manifest["notifications"]["PRE_UPGRADE"]:
-            app["notifications"]["PRE_UPGRADE"] = _filter_and_hydrate_notifications(
-                manifest["notifications"]["PRE_UPGRADE"],
-                app["current_version"],
-                app["settings"],
-            )
-        del app["settings"]
-        shutil.rmtree(extracted_app_folder)
-
-    return upgradable_apps
-
-
-#
-# ############################### #
-#        Small utilities          #
-# ############################### #
-#
-
-
-def _is_installed(app: str) -> bool:
-    return os.path.isdir(APPS_SETTING_PATH + app)
-
-
-def _assert_is_installed(app: str) -> None:
-    if not _is_installed(app):
-        raise YunohostValidationError(
-            "app_not_installed", app=app, all_apps=_get_all_installed_apps_id()
-        )
-
-
-def _installed_apps() -> List[str]:
-    return os.listdir(APPS_SETTING_PATH)
-
-
-def _get_all_installed_apps_id() -> str:
-    """
-    Return something like:
-       ' * app1
-         * app2
-         * ...'
-    """
-
-    all_apps_ids = sorted(_installed_apps())
-
-    all_apps_ids_formatted = "\n * ".join(all_apps_ids)
-    all_apps_ids_formatted = "\n * " + all_apps_ids_formatted
-
-    return all_apps_ids_formatted
-
-
-def _check_manifest_requirements(
-    manifest: Dict, action: str = ""
-) -> Iterator[Tuple[str, bool, dict[str, Any], str]]:
-    """Check if required packages are met from the manifest"""
-
-    app_id = manifest["id"]
-    logger.debug(m18n.n("app_requirements_checking", app=app_id))
-
-    # Packaging format
-    if manifest["packaging_format"] not in [1, 2]:
-        raise YunohostValidationError("app_packaging_format_not_supported")
-
-    # Yunohost version
-    required_yunohost_version = (
-        manifest["integration"].get("yunohost", "4.3").strip(">= ")
-    )
-    current_yunohost_version = get_ynh_package_version("yunohost")["version"]
-
-    yield (
-        "required_yunohost_version",
-        version.parse(required_yunohost_version)
-        <= version.parse(current_yunohost_version),
-        {"current": current_yunohost_version, "required": required_yunohost_version},
-        "app_yunohost_version_not_supported",  # i18n: app_yunohost_version_not_supported
-    )
-
-    # Architectures
-    arch_requirement = manifest["integration"]["architectures"]
-    arch = system_arch()
-
-    yield (
-        "arch",
-        arch_requirement in ["all", "?"] or arch in arch_requirement,
-        {"current": arch, "required": ", ".join(arch_requirement)},
-        "app_arch_not_supported",  # i18n: app_arch_not_supported
-    )
-
-    # Multi-instance
-    if action == "install":
-        multi_instance = manifest["integration"]["multi_instance"] is True
-        if not multi_instance:
-            apps = _installed_apps()
-            sibling_apps = [
-                a for a in apps if a == app_id or a.startswith(f"{app_id}__")
-            ]
-            multi_instance = len(sibling_apps) == 0
-
-        yield (
-            "install",
-            multi_instance,
-            {"app": app_id},
-            "app_already_installed",  # i18n: app_already_installed
-        )
-
-    # Disk
-    if action == "install":
-        root_free_space = free_space_in_directory("/")
-        var_free_space = free_space_in_directory("/var")
-        if manifest["integration"]["disk"] == "?":
-            has_enough_disk = True
-        else:
-            disk_req_bin = human_to_binary(manifest["integration"]["disk"])
-            has_enough_disk = (
-                root_free_space > disk_req_bin and var_free_space > disk_req_bin
-            )
-        free_space = binary_to_human(min(root_free_space, var_free_space))
-
-        yield (
-            "disk",
-            has_enough_disk,
-            {"current": free_space, "required": manifest["integration"]["disk"]},
-            "app_not_enough_disk",  # i18n: app_not_enough_disk
-        )
-
-    # Ram
-    ram_requirement = manifest["integration"]["ram"]
-    ram, swap = ram_available()
-    # Is "include_swap" really useful ? We should probably decide wether to always include it or not instead
-    if ram_requirement.get("include_swap", False):
-        ram += swap
-
-    if ram_requirement["build"] == "?" or ram > human_to_binary(
-        ram_requirement["build"]
-    ):
-        can_build = True
-    # When upgrading, compare the available ram to (build - runtime), because the app is already running
-    elif (
-        action == "upgrade"
-        and ram_requirement["runtime"] != "?"
-        and ram
-        > human_to_binary(ram_requirement["build"])
-        - human_to_binary(ram_requirement["runtime"])
-    ):
-        can_build = True
-    else:
-        can_build = False
-
-    # Before upgrading, the application is probably already running,
-    # and RAM rarely increases significantly from one version to the next.
-    can_run = (
-        ram_requirement["runtime"] == "?"
-        or ram > human_to_binary(ram_requirement["runtime"])
-        or action == "upgrade"
-    )
-
-    # Some apps have a higher runtime value than build ...
-    if ram_requirement["build"] != "?" and ram_requirement["runtime"] != "?":
-        max_build_runtime = (
-            ram_requirement["build"]
-            if human_to_binary(ram_requirement["build"])
-            > human_to_binary(ram_requirement["runtime"])
-            else ram_requirement["runtime"]
-        )
-    elif ram_requirement["build"] != "?":
-        max_build_runtime = ram_requirement["build"]
-    else:
-        max_build_runtime = ram_requirement["runtime"]
-
-    yield (
-        "ram",
-        can_build and can_run,
-        {"current": binary_to_human(ram), "required": max_build_runtime},
-        "app_not_enough_ram",  # i18n: app_not_enough_ram
-    )
-
-
-def _guess_webapp_path_requirement(app_folder: str) -> str:
-    # If there's only one "domain" and "path", validate that domain/path
-    # is an available url and normalize the path.
-
-    manifest = _get_manifest_of_app(app_folder)
-    raw_questions = manifest["install"]
-
-    domain_questions = [
-        question
-        for question in raw_questions.values()
-        if question.get("type") == "domain"
-    ]
-    path_questions = [
-        question
-        for question in raw_questions.values()
-        if question.get("type") == "path"
-    ]
-
-    if len(domain_questions) == 0 and len(path_questions) == 0:
-        return ""
-    if len(domain_questions) == 1 and len(path_questions) == 1:
-        return "domain_and_path"
-    if len(domain_questions) == 1 and len(path_questions) == 0:
-        if manifest.get("packaging_format", 0) < 2:
-            # This is likely to be a full-domain app...
-
-            # Confirm that this is a full-domain app This should cover most cases
-            # ...  though anyway the proper solution is to implement some mechanism
-            # in the manifest for app to declare that they require a full domain
-            # (among other thing) so that we can dynamically check/display this
-            # requirement on the webadmin form and not miserably fail at submit time
-
-            # Full-domain apps typically declare something like path_url="/" or path=/
-            # and use ynh_webpath_register or yunohost_app_checkurl inside the install script
-            install_script_content = read_file(
-                os.path.join(app_folder, "scripts/install")
-            )
-
-            if re.search(
-                r"\npath(_url)?=[\"']?/[\"']?", install_script_content
-            ) and re.search(r"ynh_webpath_register", install_script_content):
-                return "full_domain"
-
-        else:
-            # For packaging v2 apps, check if there's a permission with url being a string
-            perm_resource = manifest.get("resources", {}).get("permissions")
-            if perm_resource is not None and isinstance(
-                perm_resource.get("main", {}).get("url"), str
-            ):
-                return "full_domain"
-
-    return "?"
-
-
-def _validate_webpath_requirement(
-    args: Dict[str, Any], path_requirement: str, ignore_app: str | None = None
-) -> None:
-    domain = args.get("domain")
-    path = args.get("path")
-
-    if not domain and not path:
-        return None
-
-    if path_requirement == "domain_and_path":
-        assert domain and path
-        _assert_no_conflicting_apps(domain, path, ignore_app=ignore_app)
-
-    elif path_requirement == "full_domain":
-        assert domain
-        _assert_no_conflicting_apps(
-            domain, "/", full_domain=True, ignore_app=ignore_app
-        )
-
-
-def _get_conflicting_apps(
-    domain: str, path: str, ignore_app: str | None = None
-) -> list[tuple[str, str, str]]:
-    """
-    Return a list of all conflicting apps with a domain/path (it can be empty)
-
-    Keyword argument:
-        domain -- The domain for the web path (e.g. your.domain.tld)
-        path -- The path to check (e.g. /coffee)
-        ignore_app -- An optional app id to ignore (c.f. the change_url usecase)
-    """
-
-    from .domain import _assert_domain_exists
-    from .utils.form import DomainOption, WebPathOption
-
-    domain = DomainOption.normalize(domain)
-    path = WebPathOption.normalize(path)
-
-    # Abort if domain is unknown
-    _assert_domain_exists(domain)
-
-    # Fetch apps map
-    apps_map = app_map(raw=True)
-
-    # Loop through all apps to check if path is taken by one of them
-    conflicts = []
-    if domain in apps_map:
-        # Loop through apps
-        for p, a in apps_map[domain].items():
-            if a["id"] == ignore_app:
-                continue
-            if path == p or (
-                not (path.startswith("/.well-known/") or p.startswith("/.well-known/"))
-                and (path == "/" or p == "/")
-            ):
-                conflicts.append((p, a["id"], a["label"]))
-
-    return conflicts
-
-
-def _assert_no_conflicting_apps(
-    domain: str, path: str, ignore_app: str | None = None, full_domain: bool = False
-) -> None:
-    conflicts = _get_conflicting_apps(domain, path, ignore_app)
-
-    if conflicts:
-        apps = []
-        for path, app_id, app_label in conflicts:
-            apps.append(f" * {domain}{path} → {app_label} ({app_id})")
-
-        if full_domain:
-            raise YunohostValidationError("app_full_domain_unavailable", domain=domain)
-        else:
-            raise YunohostValidationError(
-                "app_location_unavailable", apps="\n".join(apps)
-            )
-
-
-def _make_environment_for_app_script(
-    app,
-    args={},
-    args_prefix="APP_ARG_",
-    workdir=None,
-    action=None,
-    force_include_app_settings=False,
-) -> dict[str, str]:
-    app_setting_path = os.path.join(APPS_SETTING_PATH, app)
-
-    manifest = _get_manifest_of_app(workdir if workdir else app_setting_path)
-
-    app_id, app_instance_nb = _parse_app_instance_name(app)
-
-    env_dict = {
-        "YNH_DEFAULT_PHP_VERSION": "8.2",
-        "YNH_APP_ID": app_id,
-        "YNH_APP_INSTANCE_NAME": app,
-        "YNH_APP_INSTANCE_NUMBER": str(app_instance_nb),
-        "YNH_APP_MANIFEST_VERSION": manifest.get("version", "?"),
-        "YNH_APP_PACKAGING_FORMAT": str(manifest["packaging_format"]),
-        "YNH_HELPERS_VERSION": str(
-            manifest.get("integration", {}).get("helpers_version")
-            or manifest["packaging_format"]
-        ).replace(".0", ""),
-        "YNH_ARCH": system_arch(),
-        "YNH_DEBIAN_VERSION": debian_version(),
-    }
-
-    if workdir:
-        env_dict["YNH_APP_BASEDIR"] = workdir
-
-    if action:
-        env_dict["YNH_APP_ACTION"] = action
-
-    for arg_name, arg_value in args.items():
-        arg_name_upper = arg_name.upper()
-        env_dict[f"YNH_{args_prefix}{arg_name_upper}"] = str(arg_value)
-
-    # If packaging format v2, load all settings
-    if manifest["packaging_format"] >= 2 or force_include_app_settings:
-        env_dict["app"] = app
-        data_to_redact = []
-        prefixes_or_suffixes_to_redact = [
-            "pwd",
-            "pass",
-            "passwd",
-            "password",
-            "passphrase",
-            "secret",
-            "key",
-            "token",
-        ]
-
-        for setting_name, setting_value in _get_app_settings(app).items():
-            # Ignore special internal settings like checksum__
-            # (not a huge deal to load them but idk...)
-            if setting_name.startswith("checksum__"):
-                continue
-
-            setting_value = str(setting_value)
-            env_dict[setting_name] = setting_value
-
-            # Check if we should redact this setting value
-            # (the check on the setting length exists to prevent stupid stuff like redacting empty string or something which is actually just 0/1, true/false, ...
-            if len(setting_value) > 6 and any(
-                setting_name.startswith(p) or setting_name.endswith(p)
-                for p in prefixes_or_suffixes_to_redact
-            ):
-                data_to_redact.append(setting_value)
-
-        # Special weird case for backward compatibility...
-        # 'path' was loaded into 'path_url' .....
-        if "path" in env_dict:
-            env_dict["path_url"] = env_dict["path"]
-
-        for operation_logger in OperationLogger._instances:
-            operation_logger.data_to_redact.extend(data_to_redact)
-
-    return env_dict
-
-
-def _parse_app_instance_name(app_instance_name: str) -> Tuple[str, int]:
-    """
-    Parse a Yunohost app instance name and extracts the original appid
-    and the application instance number
-
-    'yolo'      -> ('yolo', 1)
-    'yolo1'     -> ('yolo1', 1)
-    'yolo__0'   -> ('yolo__0', 1)
-    'yolo__1'   -> ('yolo', 1)
-    'yolo__23'  -> ('yolo', 23)
-    'yolo__42__72'    -> ('yolo__42', 72)
-    'yolo__23qdqsd'   -> ('yolo__23qdqsd', 1)
-    'yolo__23qdqsd56' -> ('yolo__23qdqsd56', 1)
-    """
-    match = re_app_instance_name.match(app_instance_name)
-    assert match, f"Could not parse app instance name : {app_instance_name}"
-    appid = match.groupdict().get("appid")
-    app_instance_nb_ = match.groupdict().get("appinstancenb") or "1"
-    if not appid:
-        raise Exception(f"Could not parse app instance name : {app_instance_name}")
-    if not str(app_instance_nb_).isdigit():
-        raise Exception(f"Could not parse app instance name : {app_instance_name}")
-    else:
-        app_instance_nb = int(str(app_instance_nb_))
-
-    return (appid, app_instance_nb)
-
-
-def _next_instance_number_for_app(app: str) -> int:
-    # Get list of sibling apps, such as {app}, {app}__2, {app}__4
-    apps = _installed_apps()
-    sibling_app_ids = [a for a in apps if a == app or a.startswith(f"{app}__")]
-
-    # Find the list of ids, such as [1, 2, 4]
-    sibling_ids = [_parse_app_instance_name(a)[1] for a in sibling_app_ids]
-
-    # Find the first 'i' that's not in the sibling_ids list already
-    i = 1
-    while True:
-        if i not in sibling_ids:
-            return i
-        else:
-            i += 1
-
-
-def _make_tmp_workdir_for_app(app: str | None = None) -> str:
-    # Create parent dir if it doesn't exists yet
-    if not os.path.exists(APP_TMP_WORKDIRS):
-        os.makedirs(APP_TMP_WORKDIRS)
-
-    now = int(time.time())
-
-    # Cleanup old dirs (if any)
-    for dir_ in os.listdir(APP_TMP_WORKDIRS):
-        path = os.path.join(APP_TMP_WORKDIRS, dir_)
-        # We only delete folders older than an arbitary 12 hours
-        # This is to cover the stupid case of upgrades
-        # Where many app will call 'yunohost backup create'
-        # from the upgrade script itself,
-        # which will also call this function while the upgrade
-        # script itself is running in one of those dir...
-        # It could be that there are other edge cases
-        # such as app-install-during-app-install
-        if os.stat(path).st_mtime < now - 12 * 3600:
-            shutil.rmtree(path)
-    tmpdir = tempfile.mkdtemp(prefix="app_", dir=APP_TMP_WORKDIRS)
-
-    # Copy existing app scripts, conf, ... if an app arg was provided
-    if app:
-        os.system(f"cp -a {APPS_SETTING_PATH}/{app}/* {tmpdir}")
-
-    return tmpdir
-
-
-def unstable_apps() -> list[str]:
-    output = []
-    deprecated_apps = ["mailman", "ffsync"]
-
-    for infos in app_list(full=True)["apps"]:
-        if (
-            not infos.get("from_catalog")
-            or infos.get("from_catalog", {}).get("state")
-            in [
-                "inprogress",
-                "notworking",
-            ]
-            or infos["id"] in deprecated_apps
-        ):
-            output.append(infos["id"])
-
-    return output
-
-
-def _assert_system_is_sane_for_app(manifest: AppManifest, when: Literal["pre", "post"]):
-    from .service import service_status
-
-    logger.debug("Checking that required services are up and running...")
-
-    # FIXME: in the past we had more elaborate checks about mariadb/php/postfix
-    # though they werent very formalized. Ideally we should rework this in the
-    # context of packaging v2, which implies deriving what services are
-    # relevant to check from the manifst
-
-    services = ["nginx", "fail2ban"]
-
-    # Wait if a service is reloading
-    test_nb = 0
-
-    while test_nb < 16:
-        if not any(s for s in services if service_status(s)["status"] == "reloading"):
-            break
-        time.sleep(0.5)
-        test_nb += 1
-
-    # List services currently down and raise an exception if any are found
-    services_status = {s: service_status(s) for s in services}
-    faulty_services = [
-        f"{s} ({status['status']})"
-        for s, status in services_status.items()
-        if status["status"] != "running"
-    ]
-
-    if faulty_services:
-        if when == "pre":
-            raise YunohostValidationError(
-                "app_action_cannot_be_ran_because_required_services_down",
-                services=", ".join(faulty_services),
-            )
-        elif when == "post":
-            raise YunohostError(
-                "app_action_broke_system", services=", ".join(faulty_services)
-            )
-
-    if dpkg_is_broken():
-        if when == "pre":
-            raise YunohostValidationError("dpkg_is_broken")
-        elif when == "post":
-            raise YunohostError("this_action_broke_dpkg")
-
-
 @is_flash_unit_operation()
 def app_dismiss_notification(app: str, name: Literal["post_install", "post_upgrade"]):
     assert isinstance(name, str)
@@ -3613,107 +2348,6 @@ def app_dismiss_notification(app: str, name: Literal["post_install", "post_upgra
     _assert_is_installed(app)
 
     app_setting(app, f"_dismiss_notification_{name_}", value="1")
-
-
-def _notification_is_dismissed(name, settings):
-    # Check for _dismiss_notiication_$name setting and also auto-dismiss
-    # notifications after one week (otherwise people using mostly CLI would
-    # never really dismiss the notification and it would be displayed forever)
-
-    if name == "POST_INSTALL":
-        return (
-            settings.get("_dismiss_notification_post_install")
-            or (int(time.time()) - settings.get("install_time", 0)) / (24 * 3600) > 7
-        )
-    elif name == "POST_UPGRADE":
-        # Check on update_time also implicitly prevent the post_upgrade notification
-        # from being displayed after install, because update_time is only set during upgrade
-        return (
-            settings.get("_dismiss_notification_post_upgrade")
-            or (int(time.time()) - settings.get("update_time", 0)) / (24 * 3600) > 7
-        )
-    else:
-        return False
-
-
-def _filter_and_hydrate_notifications(
-    notifications, current_version=None, data={}
-) -> dict[str, str]:
-    def is_version_more_recent_than_current_version(name, current_version):
-        current_version = str(current_version)
-        return _parse_app_version(name) > _parse_app_version(current_version)
-
-    out = {
-        # Should we render the markdown maybe? idk
-        name: _hydrate_app_template(_value_for_locale(content_per_lang), data)
-        for name, content_per_lang in notifications.items()
-        if current_version is None
-        or name == "main"
-        or is_version_more_recent_than_current_version(name, current_version)
-    }
-
-    # Filter out empty notifications (notifications may be empty because of if blocks)
-    return {
-        name: content for name, content in out.items() if content and content.strip()
-    }
-
-
-def _display_notifications(notifications: dict[str, str], force=False) -> None:
-    if not notifications:
-        return
-
-    for name, content in notifications.items():
-        print("==========")
-        print(content)
-    print("==========")
-
-    # i18n: confirm_notifications_read
-    _ask_confirmation("confirm_notifications_read", kind="simple", force=force)
-
-
-# FIXME: move this to Moulinette
-def _ask_confirmation(
-    question: str,
-    params: dict = {},
-    kind: Literal["simple", "soft", "hard"] = "hard",
-    force: bool = False,
-) -> None:
-    """
-    Ask confirmation
-
-    Keyword argument:
-        question -- m18n key or string
-        params -- dict of values passed to the string formating
-        kind -- "hard": ask with "Yes, I understand", "soft": "Y/N", "simple": "press enter"
-        force -- Will not ask for confirmation
-
-    """
-    if force or Moulinette.interface.type == "api":
-        return
-
-    # If ran from the CLI in a non-interactive context,
-    # skip confirmation (except in hard mode)
-    if not os.isatty(1) and kind in ["simple", "soft"]:
-        return
-    if kind == "simple":
-        answer = Moulinette.prompt(
-            m18n.n(question, answers="Press enter to continue", **params),
-            color="yellow",
-        )
-        answer = True
-    elif kind == "soft":
-        answer = Moulinette.prompt(
-            m18n.n(question, answers="Y/N", **params), color="yellow"
-        )
-        answer = answer.upper() == "Y"
-    else:
-        answer = Moulinette.prompt(
-            m18n.n(question, answers="Yes, I understand", **params), color="red"
-        )
-        answer = answer == "Yes, I understand"
-
-    if not answer:
-        raise YunohostError("aborting")
 
 
 def regen_mail_app_user_config_for_dovecot_and_postfix(
